requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
bleach==5.0.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
django-ipware==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.9.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.15.0
MarkupSafe==2.1.1
packaging==21.3
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
